refactor: QuotaCheckerImplTests
