+++ Modules/Quotas/test/Quotas.Application.Tests/Tests/QuotaCheck/UserContextStub.cs
@@ -1,0 +1,51 @@
+﻿using Enmeshed.BuildingBlocks.Application.Abstractions.Infrastructure.UserContext;
+using Enmeshed.DevelopmentKit.Identity.ValueObjects;
+
+namespace Backbone.Modules.Quotas.Application.Tests.Tests.QuotaCheck;
+
+internal class UserContextStub : IUserContext
+{
+    public UserContextStub()
+    {
+    }
+
+    public IdentityAddress GetAddress()
+    {
+        return IdentityAddress.Create(new byte[]{0}, "id1");
+    }
+
+    public IdentityAddress GetAddressOrNull()
+    {
+        throw new NotImplementedException();
+    }
+
+    public DeviceId GetDeviceId()
+    {
+        throw new NotImplementedException();
+    }
+
+    public DeviceId GetDeviceIdOrNull()
+    {
+        throw new NotImplementedException();
+    }
+
+    public IEnumerable<string> GetRoles()
+    {
+        throw new NotImplementedException();
+    }
+
+    public SubscriptionPlan GetSubscriptionPlan()
+    {
+        throw new NotImplementedException();
+    }
+
+    public string GetUserId()
+    {
+        throw new NotImplementedException();
+    }
+
+    public string GetUserIdOrNull()
+    {
+        throw new NotImplementedException();
+    }
+}
